rusty: Use new topology crate
Now that we have this new Topology crate, let's update Rusty to use it
instead of using the old one.
